refactor(fiat-exchange): remove unused component
